# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Rewrd API",
    "version": "1.0.0",
...
    "contact": {
      "name": "Rewrd Engineering",
      "url": "https://getrewrd.com"
    }
  },
  "servers": [
    {
      "url": "https://rewrd-api-staging.fly.dev/v1",
      "description": "Development Server"
    },
    {
      "url": "http://localhost:3004/v1",
      "description": "Local Server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your API Key (e.g. sk_test_...)"
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "description": "Standard error response returned by all endpoints on failure.",
        "properties": {
          "status": {
            "type": "boolean",
            "example": false,
            "description": "Always `false` for error responses."
          },
          "message": {
            "type": "string",
            "example": "Customer not found",
            "description": "Human-readable error message."
          },
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "example": "not_found",
                "description": "Machine-readable error code for programmatic handling."
              },
              "request_id": {
                "type": "string",
                "example": "a1b2c3d4-e5f6-7890-abcd-ef1234567890",
                "description": "Unique identifier for this request, useful for support inquiries."
              },
              "description": {
                "type": "string",
                "example": "Customer not found",
                "description": "Detailed description of the error."
              }
            }
          }
        }
      },
...
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
    "/customers": {
      "post": {
        "summary": "Create or Update a Customer",
        "operationId": "createOrUpdateCustomer",
        "description": "Creates a new customer or updates an existing one. This is an **upsert** operation — meaning \"update if exists, insert (create) if not\".\n\n**How the matching works:**\nRewrd uses the `phone_number` to check if a customer already exists in your merchant account:\n- **New customer** (phone number not found) → A new customer record is created with status `active` and 0 points balance.\n- **Existing customer** (phone number already exists) → Only the fields you provide are updated. Fields you don't include are left unchanged.\n\nThis makes it safe to call this endpoint every time a user signs up or logs in — you won't create duplicates.\n\n---\n\n**Quick Start — Enroll your first customer:**\n\n```bash\ncurl -X POST https://api.getrewrd.com/v1/customers \\\n  -H \"Authorization: Bearer sk_test_your_api_key\" \\\n  -H \"Content-Type: application/json\" \\\n  -H \"Idempotency-Key: enroll-jane-001\" \\\n  -d '{\n    \"customer_email\": \"jane@example.com\",\n    \"phone_number\": \"+2348012345678\"\n  }'\n```\n\n---\n\nThis endpoint requires an `Idempotency-Key` header to prevent duplicate customer creation from retried requests.\n",
        "tags": [
          "Customers"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[a-zA-Z0-9_-]{1,100}$"
            },
            "description": "Unique key for idempotent request handling. Must be alphanumeric with hyphens or underscores, 1–100 characters.",
            "example": "customer-create-20260203-001"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "phone_number",
                  "customer_email"
                ],
                "properties": {
                  "customer_email": {
                    "type": "string",
                    "format": "email",
                    "description": "Customer's email address.",
                    "example": "jane@example.com"
                  },
                  "phone_number": {
                    "type": "string",
                    "description": "Customer's phone number. Used as the unique identifier for matching.",
                    "example": "+2348012345678"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Customer created or updated successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Customer created successfully"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Customer"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error — missing required fields or invalid `Idempotency-Key`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized — missing or invalid API key.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "List Customers",
        "operationId": "listCustomers",
        "description": "Retrieves a paginated list of all customers belonging to your merchant account.\n\nYou can filter results by `email` or `phone_number` to quickly find specific customers. Results are sorted by creation date (newest first).\n",
        "tags": [
          "Customers"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            },
            "description": "Page number for pagination (1-indexed).",
            "example": 1
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 50
            },
            "description": "Number of customers to return per page (max 100).",
            "example": 20
          },
          {
            "in": "query",
            "name": "email",
            "schema": {
              "type": "string"
            },
            "description": "Filter customers by exact email address match.",
            "example": "jane@example.com"
          },
          {
            "in": "query",
            "name": "phone_number",
            "schema": {
              "type": "string"
            },
            "description": "Filter customers by exact phone number match.",
            "example": "+2348012345678"
          }
        ],
        "responses": {
          "200": {
            "description": "List of customers retrieved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Customers retrieved successfully"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Customer"
                      }
                    },
                    "pagination": {
                      "$ref": "#/components/schemas/Pagination"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized — missing or invalid API key.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/customers/{uid}": {
      "get": {
        "summary": "Get a Customer",
        "operationId": "getCustomer",
        "description": "Retrieves the full details of a single customer by their unique identifier (`uid`).\n\nUse this endpoint to check a customer's current points balance, status, and profile information before performing point operations.\n",
        "tags": [
          "Customers"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "uid",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The unique identifier of the customer.",
            "example": "26833813"
          }
        ],
        "responses": {
          "200": {
            "description": "Customer details retrieved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Customer fetched successfully"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Customer"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized — missing or invalid API key.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Customer not found — no customer exists with the given UID.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update a Customer",
        "operationId": "updateCustomer",
        "description": "Updates the profile information of an existing customer. Only the fields provided in the request body are updated; omitted fields are left unchanged.\n\n**Note:** The customer must have an `active` status to be updated. Restricted customers must be unrestricted first.\n",
        "tags": [
          "Customers"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "uid",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The unique identifier of the customer to update.",
            "example": "26833813"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "description": "Updated email address.",
                    "example": "jane.new@example.com"
                  },
                  "phone_number": {
                    "type": "string",
                    "description": "Updated phone number.",
                    "example": "+2349087654321"
                  },
                  "first_name": {
                    "type": "string",
                    "description": "Updated first name.",
                    "example": "Janet"
                  },
                  "last_name": {
                    "type": "string",
                    "description": "Updated last name.",
                    "example": "Smith"
                  },
                  "date_of_birth": {
                    "type": "string",
                    "description": "Updated date of birth in ISO 8601 format.",
                    "example": "1995-06-15"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Customer updated successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Customer updated successfully"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Customer"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized — missing or invalid API key.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Customer is not active — restricted customers cannot be updated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Customer not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/customers/{uid}/restrict": {
      "patch": {
        "summary": "Restrict a Customer",
        "operationId": "restrictCustomer",
        "description": "Restricts a customer, preventing them from earning or redeeming points. The customer's existing points balance is preserved but frozen.\n\n**When to use this:**\n- **Fraud prevention** — You suspect a customer is abusing the program\n- **Dispute resolution** — A customer disputes a transaction and you want to freeze their account while investigating\n- **Temporary suspension** — The customer requested a pause on their account\n\n**Customer status lifecycle:**\n```\nactive ──(restrict)──► restricted ──(unrestrict)──► active\n```\n\nOnly `active` customers can be restricted. If the customer is already `restricted`, this endpoint will return a `403` error.\n",
        "tags": [
          "Customers"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "uid",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The unique identifier of the customer to restrict.",
            "example": "26833813"
          }
        ],
        "responses": {
          "200": {
            "description": "Customer restricted successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Customer restricted successfully"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized — missing or invalid API key.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Customer is not active — cannot restrict an already restricted or inactive customer.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Customer not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}