# src/routes/customerRoutes.ts
import { Router } from "express";
import { requireIdempotency } from "../middleware/idempotency";
import { requireCustomer } from "../middleware/requireCustomer";
import { validateRequest } from "../middleware/validateRequest";
import { requireActiveCustomer } from "../middleware/requireActiveCustomer";
import { createCustomerSchema, getCustomerSchema, listCustomersSchema, updateCustomerSchema } from "../schema/customerSchema";
import { createOrUpdateCustomer, getCustomer, listCustomers, updateCustomer, restrictCustomer, unrestrictCustomer } from "../controllers/customerController";

const router = Router();

/**
 * @swagger
 * components:
 *   schemas:
 *     Customer:
 *       type: object
 *       description: Represents a customer enrolled in your loyalty program.
 *       properties:
 *         uid:
 *           type: string
 *           description: Unique identifier for the customer, generated by Rewrd.
 *           example: "26833813"
 *         email:
 *           type: string
 *           format: email
 *           nullable: true
 *           description: Customer's email address. Used for notifications if configured.
 *           example: "jane@example.com"
 *         phone_number:
 *           type: string
 *           description: Customer's phone number. This is the primary identifier used for matching.
 *           example: "+2348012345678"
 *         first_name:
 *           type: string
 *           nullable: true
 *           description: Customer's first name.
 *           example: "Jane"
 *         last_name:
 *           type: string
 *           nullable: true
 *           description: Customer's last name.
 *           example: "Doe"
 *         date_of_birth:
 *           type: string
 *           format: date
 *           nullable: true
 *           description: Customer's date of birth in ISO 8601 format.
 *           example: "1995-06-15"
 *         status:
 *           type: string
 *           enum: [active, inactive, restricted]
 *           description: "Current status of the customer. `active`: can earn and redeem points. `restricted`: blocked from all point operations."
 *           example: "active"
 *         points_balance:
 *           type: number
 *           description: Current available points balance for this customer.
 *           example: 1500
 *         created_at:
 *           type: string
 *           format: date-time
 *           description: Timestamp when the customer was first created.
 *           example: "2026-01-15T10:30:00.000Z"
 */

/**
 * @swagger
 * /customers:
 *   post:
 *     summary: Create or Update a Customer
 *     operationId: createOrUpdateCustomer
 *     description: |
 *       Creates a new customer or updates an existing one. This is an **upsert** operation — meaning "update if exists, insert (create) if not".
 *
 *       **How the matching works:**
 *       Rewrd uses the `phone_number` to check if a customer already exists in your merchant account:
 *       - **New customer** (phone number not found) → A new customer record is created with status `active` and 0 points balance.
 *       - **Existing customer** (phone number already exists) → Only the fields you provide are updated. Fields you don't include are left unchanged.
 *
 *       This makes it safe to call this endpoint every time a user signs up or logs in — you won't create duplicates.
 *
 *       ---
 *
 *       **Quick Start — Enroll your first customer:**
 *
 *       ```bash
 *       curl -X POST https://api.getrewrd.com/v1/customers \
 *         -H "Authorization: Bearer sk_test_your_api_key" \
 *         -H "Content-Type: application/json" \
 *         -H "Idempotency-Key: enroll-jane-001" \
 *         -d '{
 *           "customer_email": "jane@example.com",
 *           "phone_number": "+2348012345678"
 *         }'
 *       ```
 *
 *       ---
 *
 *       This endpoint requires an `Idempotency-Key` header to prevent duplicate customer creation from retried requests.
 *     tags: [Customers]
 *     security:
 *       - BearerAuth: []
 *     parameters:
 *       - in: header
 *         name: Idempotency-Key
 *         required: true
 *         schema:
 *           type: string
 *           pattern: '^[a-zA-Z0-9_-]{1,100}$'
 *         description: Unique key for idempotent request handling. Must be alphanumeric with hyphens or underscores, 1–100 characters.
 *         example: "customer-create-20260203-001"
 *     requestBody:
 *       required: true
 *       content:
 *         application/json:
 *           schema:
 *             type: object
 *             required:
 *               - phone_number
 *               - customer_email
 *             properties:
 *               customer_email:
 *                 type: string
 *                 format: email
 *                 description: Customer's email address.
 *                 example: "jane@example.com"
 *               phone_number:
 *                 type: string
 *                 description: Customer's phone number. Used as the unique identifier for matching.
 *                 example: "+2348012345678"
 *     responses:
 *       200:
 *         description: Customer created or updated successfully.
 *         content:
 *           application/json:
 *             schema:
 *               type: object
 *               properties:
 *                 status:
 *                   type: boolean
 *                   example: true
 *                 message:
 *                   type: string
 *                   example: "Customer created successfully"
 *                 data:
 *                   $ref: '#/components/schemas/Customer'
 *       400:
 *         description: Validation error — missing required fields or invalid `Idempotency-Key`.
 *         content:
 *           application/json:
 *             schema:
 *               $ref: '#/components/schemas/ErrorResponse'
 *       401:
 *         description: Unauthorized — missing or invalid API key.
 *         content:
 *           application/json:
 *             schema:
 *               $ref: '#/components/schemas/ErrorResponse'
 */
router.post("/", requireIdempotency, validateRequest(createCustomerSchema), createOrUpdateCustomer);

/**
 * @swagger
 * /customers:
 *   get:
 *     summary: List Customers
 *     operationId: listCustomers
 *     description: |
 *       Retrieves a paginated list of all customers belonging to your merchant account.
 *
 *       You can filter results by `email` or `phone_number` to quickly find specific customers. Results are sorted by creation date (newest first).
 *     tags: [Customers]
 *     security:
 *       - BearerAuth: []
 *     parameters:
 *       - in: query
 *         name: page
 *         schema:
 *           type: integer
 *           default: 1
 *         description: Page number for pagination (1-indexed).
 *         example: 1
 *       - in: query
 *         name: limit
 *         schema:
 *           type: integer
 *           default: 50
 *         description: Number of customers to return per page (max 100).
 *         example: 20
 *       - in: query
 *         name: email
 *         schema:
 *           type: string
 *         description: Filter customers by exact email address match.
 *         example: "jane@example.com"
 *       - in: query
 *         name: phone_number
 *         schema:
 *           type: string
 *         description: Filter customers by exact phone number match.
 *         example: "+2348012345678"
 *     responses:
 *       200:
 *         description: List of customers retrieved successfully.
 *         content:
 *           application/json:
 *             schema:
 *               type: object
 *               properties:
 *                 status:
 *                   type: boolean
 *                   example: true
 *                 message:
 *                   type: string
 *                   example: "Customers retrieved successfully"
 *                 data:
 *                   type: array
 *                   items:
 *                     $ref: '#/components/schemas/Customer'
 *                 pagination:
 *                   $ref: '#/components/schemas/Pagination'
 *       401:
 *         description: Unauthorized — missing or invalid API key.
 *         content:
 *           application/json:
 *             schema:
 *               $ref: '#/components/schemas/ErrorResponse'
 */
router.get("/", validateRequest(listCustomersSchema), listCustomers);

/**
 * @swagger
 * /customers/{uid}:
 *   get:
 *     summary: Get a Customer
 *     operationId: getCustomer
 *     description: |
 *       Retrieves the full details of a single customer by their unique identifier (`uid`).
 *
 *       Use this endpoint to check a customer's current points balance, status, and profile information before performing point operations.
 *     tags: [Customers]
 *     security:
 *       - BearerAuth: []
 *     parameters:
 *       - in: path
 *         name: uid
 *         required: true
 *         schema:
 *           type: string
 *         description: The unique identifier of the customer.
 *         example: "26833813"
 *     responses:
 *       200:
 *         description: Customer details retrieved successfully.
 *         content:
 *           application/json:
 *             schema:
 *               type: object
 *               properties:
 *                 status:
 *                   type: boolean
 *                   example: true
 *                 message:
 *                   type: string
 *                   example: "Customer fetched successfully"
 *                 data:
 *                   $ref: '#/components/schemas/Customer'
 *       404:
 *         description: Customer not found — no customer exists with the given UID.
 *         content:
 *           application/json:
 *             schema:
 *               $ref: '#/components/schemas/ErrorResponse'
 *       401:
 *         description: Unauthorized — missing or invalid API key.
 *         content:
 *           application/json:
 *             schema:
 *               $ref: '#/components/schemas/ErrorResponse'
 */
router.get("/:uid", validateRequest(getCustomerSchema), requireCustomer, getCustomer);

/**
 * @swagger
 * /customers/{uid}:
 *   put:
 *     summary: Update a Customer
 *     operationId: updateCustomer
 *     description: |
 *       Updates the profile information of an existing customer. Only the fields provided in the request body are updated; omitted fields are left unchanged.
 *
 *       **Note:** The customer must have an `active` status to be updated. Restricted customers must be unrestricted first.
 *     tags: [Customers]
 *     security:
 *       - BearerAuth: []
 *     parameters:
 *       - in: path
 *         name: uid
 *         required: true
 *         schema:
 *           type: string
 *         description: The unique identifier of the customer to update.
 *         example: "26833813"
 *     requestBody:
 *       required: true
 *       content:
 *         application/json:
 *           schema:
 *             type: object
 *             properties:
 *               email:
 *                 type: string
 *                 description: Updated email address.
 *                 example: "jane.new@example.com"
 *               phone_number:
 *                 type: string
 *                 description: Updated phone number.
 *                 example: "+2349087654321"
 *               first_name:
 *                 type: string
 *                 description: Updated first name.
 *                 example: "Janet"
 *               last_name:
 *                 type: string
 *                 description: Updated last name.
 *                 example: "Smith"
 *               date_of_birth:
 *                 type: string
 *                 description: Updated date of birth in ISO 8601 format.
 *                 example: "1995-06-15"
 *     responses:
 *       200:
 *         description: Customer updated successfully.
 *         content:
 *           application/json:
 *             schema:
 *               type: object
 *               properties:
 *                 status:
 *                   type: boolean
 *                   example: true
 *                 message:
 *                   type: string
 *                   example: "Customer updated successfully"
 *                 data:
 *                   $ref: '#/components/schemas/Customer'
 *       404:
 *         description: Customer not found.
 *         content:
 *           application/json:
 *             schema:
 *               $ref: '#/components/schemas/ErrorResponse'
 *       403:
 *         description: Customer is not active — restricted customers cannot be updated.
 *         content:
 *           application/json:
 *             schema:
 *               $ref: '#/components/schemas/ErrorResponse'
 *       401:
 *         description: Unauthorized — missing or invalid API key.
 *         content:
 *           application/json:
 *             schema:
 *               $ref: '#/components/schemas/ErrorResponse'
 */
router.put("/:uid", validateRequest(updateCustomerSchema), requireCustomer, requireActiveCustomer, updateCustomer);

/**
 * @swagger
 * /customers/{uid}/restrict:
 *   patch:
 *     summary: Restrict a Customer
 *     operationId: restrictCustomer
 *     description: |
 *       Restricts a customer, preventing them from earning or redeeming points. The customer's existing points balance is preserved but frozen.
 *
 *       **When to use this:**
 *       - **Fraud prevention** — You suspect a customer is abusing the program
 *       - **Dispute resolution** — A customer disputes a transaction and you want to freeze their account while investigating
 *       - **Temporary suspension** — The customer requested a pause on their account
 *
 *       **Customer status lifecycle:**
 *       ```
 *       active ──(restrict)──► restricted ──(unrestrict)──► active
 *       ```
 *
 *       Only `active` customers can be restricted. If the customer is already `restricted`, this endpoint will return a `403` error.
 *     tags: [Customers]
 *     security:
 *       - BearerAuth: []
 *     parameters:
 *       - in: path
 *         name: uid
 *         required: true
 *         schema:
 *           type: string
 *         description: The unique identifier of the customer to restrict.
 *         example: "26833813"
 *     responses:
 *       200:
 *         description: Customer restricted successfully.
 *         content:
 *           application/json:
 *             schema:
 *               type: object
 *               properties:
 *                 status:
 *                   type: boolean
 *                   example: true
 *                 message:
 *                   type: string
 *                   example: "Customer restricted successfully"
 *       404:
 *         description: Customer not found.
 *         content:
 *           application/json:
 *             schema:
 *               $ref: '#/components/schemas/ErrorResponse'
 *       403:
 *         description: Customer is not active — cannot restrict an already restricted or inactive customer.
 *         content:
 *           application/json:
 *             schema:
 *               $ref: '#/components/schemas/ErrorResponse'
 *       401:
 *         description: Unauthorized — missing or invalid API key.
 *         content:
 *           application/json:
 *             schema:
 *               $ref: '#/components/schemas/ErrorResponse'
 */
router.patch("/:uid/restrict", validateRequest(getCustomerSchema), requireCustomer, requireActiveCustomer, restrictCustomer);

/**
 * @swagger
 * /customers/{uid}/unrestrict:
 *   patch:
 *     summary: Unrestrict a Customer
 *     operationId: unrestrictCustomer
 *     description: |
 *       Removes the restriction on a customer, restoring their ability to earn and redeem points. Their previously frozen points balance becomes available again.
 *     tags: [Customers]
 *     security:
 *       - BearerAuth: []
 *     parameters:
 *       - in: path
 *         name: uid
 *         required: true
 *         schema:
 *           type: string
 *         description: The unique identifier of the customer to unrestrict.
 *         example: "26833813"
 *     responses:
 *       200:
 *         description: Customer unrestricted successfully.
 *         content:
 *           application/json:
 *             schema:
 *               type: object
 *               properties:
 *                 status:
 *                   type: boolean
 *                   example: true
 *                 message:
 *                   type: string
 *                   example: "Customer unrestricted successfully"
 *       404:
 *         description: Customer not found.
 *         content:
 *           application/json:
 *             schema:
 *               $ref: '#/components/schemas/ErrorResponse'
 *       401:
 *         description: Unauthorized — missing or invalid API key.
 *         content:
 *           application/json:
 *             schema:
 *               $ref: '#/components/schemas/ErrorResponse'
 */
router.patch("/:uid/unrestrict", validateRequest(getCustomerSchema), requireCustomer, unrestrictCustomer);

export default router;